pyproject.toml
[project]
name = "ai-agent-whatsapp"
version = "0.1.0"
description = "AI-based WhatsApp Task Extraction & Management Agent"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.135.2",
    "langchain>=1.2.13",
    "langchain-ollama>=1.0.1",
    "langchain-openai>=1.1.12",
    "langgraph>=1.1.3",
    "langgraph-cli[inmem]>=0.4.19",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.22",
    "sqlalchemy>=2.0.48",
    "twilio>=9.10.4",
    "uvicorn>=0.42.0",
]

[dependency-groups]
dev = [
    "black>=26.3.1",
    "httpx>=0.28.1",
    "mypy>=1.19.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.15.8",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
